requirements.txt
streamlit
opencv-python-headless
#opencv-python==3.4.9.33
Pillow
